// src/app/login-api.ts
// export class LoginApi {
//     tokenType:string="";
//     accessToken:string="";
//     // constructor(tokenType:string,accessToken:string){
//     //     this.accessToken=accessToken;
//     //     this.tokenType=tokenType;
//     // }
// }

export interface  LoginApi {
    tokenType: string
    accessToken: string
  }